+++ test/nhsd_ods/organisation_test.rb
@@ -1,0 +1,28 @@
+require 'test_helper'
+require 'ndr_lookup/nhsd_ods/organisation'
+
+module NdrLookup
+  module NhsdOds
+    # The orgaisation class tests
+    class OrganisationTest < Minitest::Test
+      def organisation_should_have_model_like_behaviour
+        url  = ODS_ENDPOINT + 'organisations/X26'
+        file = File.new(RESPONSES_DIR + '/nhsd_ods/organisation_find_success_response.txt')
+        stub_request(:get, url).to_return(file)
+        org = NdrLookup::NhsdOds::Organisation.find('X26')
+
+        assert_equal org.name, 'NHS DIGITAL'
+      end
+
+      def should_raise_error_if_not_valid_find_method
+        url  = ODS_ENDPOINT + 'organisations'
+        file = File.new(RESPONSES_DIR + '/nhsd_ods/organisation_all_not_acceptable_response.txt')
+        stub_request(:get, url).to_return(file)
+
+        assert_raises do
+          NdrLookup::NhsdOds::Organisation.all
+        end
+      end
+    end
+  end
+end