+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## [Unreleased]
+* Added NHS Digital ODS API client
+
 ### Added
 * Added existing ArcGIS LocatorHub address `Rectify` client